Added hardhat and barebones deployment scripts with tasks (#5)
* Added barebones deployment scripts

The deployment scripts added are not finalized but merely allow for successful deployment against forked networks. They have not been tested against real networks. These are to be used as follows:

- Run both the deployment scripts (scripts starting with deploy) first. This can be done simultaneously.
- Run the tasks based on the output of the deployment scripts. You will need the comptroller addresses to run these tasks.

The tasks set the comptroller address in the comptrollers and transfer the ownership of the ProxyAdmin contracts to the multisigs given as an argument when the tasks are run.

* Minor changes in .env.sample and hh config

* Stagin deployment complete

* Added CheckComptroller tasks and configs

Configs in this case serve as deployment (initialization) arguments for both the Comptrollers. CheckComptroller tasks use these configs to verify that everything is initialized correctly in the
comptroller contracts.

* Added OZ Defender package and upgrade scripts

OpenZeppelin Defender allows us to modify the comptrollers in any way we would like. This involves upgrading the implementation contracts, pausing/unpausing comptrollers, modifying a numerical
parameter etc. The upgrade scripts proposes a new upgrade of implementation contracts using OZ defender library which also checks if the new implementation doesn't contain any storage slots violation.

* Fixed error in sending cross chain message

- ABI encoding of "buyBackFromL1(address,address,uint256)" is different from "buyBackFromL1(address,address,uint)" due to this, cross chain message calls weren't valid as the latter method doesn't
exist in our L2Comptroller contract.

- Also updated the upgrade scripts to compile all the contracts before creating a proposal to upgrade.

- Minor fixes and cleanup.

* Requested changes made

Changes as per [this](https://github.com/dhedge/buyback-contract/issues/8#issuecomment-1535969467) comment have been made.